Add Support for Privacy Function URL (#6)
This PR updates the SDK to support a privacy function URL configuration option. The `init` method now accepts either a project token string or a configuration object that can include an `endpoint` and a `privacyFnUrl`. When provided, the privacy function URL is added as a `data-pz-privacy-src` attribute to the injected script.

The README has been updated to document this new functionality.

Resolves ENG-6137